% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawl.R
\name{assert_blast_dbs_exist}
\alias{assert_blast_dbs_exist}
\title{Assertion that BLAST DB indicator files exist}
\usage{
assert_blast_dbs_exist(
  filenames,
  .var.name = checkmate::vname(filenames),
  add = NULL
)
}
\arguments{
\item{filenames}{[character] One or more base paths (without extension) of
the BLAST databases to check.}

\item{.var.name}{[\code{character(1)}]\cr
Name of the checked object to print in error messages. Defaults to
the heuristic implemented in checkmate's \code{vname}.}

\item{add}{[\code{checkmate::AssertCollection}]\cr
Collection to store assertions. See checkmate's \code{AssertCollection}.}
}
\value{
Invisibly returns its \code{filenames} argument if all base paths
  have at least one of \code{.nsq} or \code{.psq}; otherwise throws an error
  with an informative message.
}
\description{
Assertion wrapper generated via \code{checkmate::makeAssertionFunction()} from
\code{check_blast_dbs_exist}. This assertion will throw an error if any of
the supplied base paths are missing both \code{.nsq} and \code{.psq}.
On success it returns its argument invisibly.
}
\examples{
\dontrun{
assert_blast_dbs_exist(c("/data/db/nt", "/data/db/prot"))
}
}
