% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawl.R
\name{assert_files_exist}
\alias{assert_files_exist}
\title{Assertion function for files existing}
\usage{
assert_files_exist(
  filenames,
  .var.name = checkmate::vname(filenames),
  add = NULL
)
}
\arguments{
\item{filenames}{[character] One or more file paths to check.}

\item{.var.name}{[\code{character(1)}]\cr
Name of the checked object to print in error messages. Defaults to
the heuristic implemented in checkmate's \code{vname}.}

\item{add}{[\code{checkmate::AssertCollection}]\cr
Collection to store assertions. See checkmate's \code{AssertCollection}.}
}
\value{
Invisibly returns its \code{filenames} argument if all files exist;
  otherwise throws an error with a descriptive message.
}
\description{
Assertion wrapper produced via \code{checkmate::makeAssertionFunction()} from
\code{check_files_exist}. When used, this assertion will throw an error if
any of the supplied file paths do not exist. On success it returns its
argument invisibly.
}
\examples{
\dontrun{
assert_files_exist(c("/path/to/file1", "/path/to/file2"))
}

}
