% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawl.R
\name{read_blast_tsv}
\alias{read_blast_tsv}
\title{Read a BLAST TSV file (wrapper for readr::read_tsv)}
\usage{
read_blast_tsv(file, col_names, col_types)
}
\arguments{
\item{file}{\code{[character(1)]}\cr
Path to the TSV file.}

\item{col_names}{\code{[logical | character]}\cr
See \code{readr::read_tsv}.}

\item{col_types}{\code{[cols | NULL | character]}\cr
Column type specification passed \code{readr::read_tsv}.}
}
\value{
A tibble with the parsed TSV contents.
}
\description{
Thin wrapper around \code{readr::read_tsv} used to make file reading easier
to mock in tests.
}
\examples{
\dontrun{
read_blast_tsv("blast_output.tsv", col_names = TRUE)
}
}
\keyword{internal}
