% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawl.R
\name{sys_which}
\alias{sys_which}
\title{Find a command in PATH or a specific directory}
\usage{
sys_which(command, directory = NULL)
}
\arguments{
\item{command}{\code{[character(1)]}\cr
The command name to look for.}

\item{directory}{\code{[character(1) | NULL]}\cr
Optional directory to search in.
If non-\code{NULL}, the directory is prepended to \code{command} and the
presence of that path is tested. May be \code{NULL}.}
}
\value{
A single string: the path to the command if found, or the empty
  string (\code{""}) if not found.
}
\description{
Like \code{Sys.which} but accepts a scalar \code{command} and an optional
single \code{directory} argument. If \code{directory} is supplied the call
checks for the command in that directory only (by prepending the directory
to the command before calling \code{Sys.which}).
}
\details{
The caller must check if the returned string is empty (meaning the command
was not found).
}
\examples{
\dontrun{
sys_which("blastn")
sys_which("blastn", directory = "/usr/local/bin")
}

}
